Fix hang on mac and linux
The TRAILING_ZEROES did not ensure correct evaluation order (it behaved incorrectly when x was a ternary expression for example). To fix this I've replaced these macros with inline functions. Adding brackets around the first x in the TRAILING_ZEROES macro would have worked as well, but we're probably better off using an inline function in this case.