added resizability configuration and screen centering for DirectX and SDL windows
The option to enable user resizing of the window can be enabled or disabled upon initialization (through constructor parameters) for both DirectX and SDL windows. A new method was created for the C bindings in order to maintain compatibility.

SDL was provided with position getters and a position setter method to match DirectX's implementation. Furthermore, both window implementations were provided with a method to center the window to the screen.