window style maintained upon return from fullscreen in directx, Hidden flag added
Upon initialization of the DirectX window, the now (somewhat) customizable style is now saved into the window-specific struct, dx_events. This allows the style to be maintained whenever the window is changed to windowed mode.

The Hidden flag was also added to provide an option for the window's visibility upon initialization. More flags could be provided for consistency in comparison to the SDL implementation, but awaiting further discussion.

As a final minor change, the SDL implementation is no longer forced to use the SDL_WINDOW_SHOWN flag in non-mobile builds.