Fixed DX cursor and added ability to check cursor visibility
The DirectX's cursor now functions as desired. The cursor is now saved and reapplied upon the WM_SETCURSOR event. Furthermore, when the cursor is not shown, it will still be visible on non-client areas.

A method for checking the cursor's visibility has been added to both DX and SDL implementations.