window min/max size and opacity configuration added
Both DirectX and SDL implementations can now be provided with a minimum and maximum size for resizing. The DirectX version is based off of SDL's; as such, the limitations are based on the client size, not the entire window.

Furthermore, both implementations now have the option to change their windows' opacity. Once again DirectX's is based on SDL's for consistency.