Hide js/gamepad subsystem init latency in a thread
On some Windows 10 systems, doing this is **extremely** slow.

I'm getting tired of waiting 5-10 extra seconds on my dev machine
before getting a window and it's entirely unclear if this is an SDL
bug or a Windows bug. One theory about this (from @rcgordon) is
that it's a timeout from a non-responsive device, in which case
this is probably the right idea (or at least a sane one) anyway.

This trades the time at window creation with some delay before you
start receiving gamepad inputs, which is much more tolerable.

Also, don't initialize the audio subsystem, it's not used/exposed.