[UNIX] Don't crash on SIGPIPE
Currently, when running on linux (didn't test on macos) if remote socket hangs up the whole application silently crashes with SIGPIPE (and no stacktrace). I traced it all the way to libuv that does this on purpose for performance reasons (see https://github.com/joyent/libuv/issues/1254#issuecomment-41185781). The recommended (as I understood it) way to prevent crashes is to redirect SIGPIPE to SIG_IGN and handle socket errors in user code instead.