Add support for SDL_GetGlobalMouseState
This allows us to track drag/resize movements that go outside the window; fixing the issue where dragging too fast will lose a handle on the window.

Viewports will function without this, but moving/resizing windows sucks.

Unsure if there's a parallel to this in DX, but we can't use that backend currently anyway since it lacks single driver multi window support at the moment.