Keep makefile tmp files within hlgen.makefilepath
Previously it would always generate them next to the .c files,
regardless of whether hlgen.makefilepath had been set. Now it ensures
that it doesn't write outside of the Makefile directory.

This means that it first has to create the directories, otherwise the
compiler complains about directories not existing when it tries to
write the output files.

".SUFFIXES" also has to be set as an empty target, otherwise the built-
in rules mess up the rule dependencies.