Reduce cl.exe calls in vs hlc templates
Since the %(Filename) variable is unique for each compiled file, this
resulted in a separate cl.exe invocation for each file. This makes us
unable to make use of /MP to speed up cl.exe.

We can instead simply pass the directory and the file name will be
added to the end automatically. This still means that there is a
separate cl.exe invocation per directory, however, this is much faster
already. The %(RelativeDir) must remain otherwise all object files are
placed in a single directory. This would be a problem if two files share
a name.