[hlc] Implement automatic compilation with make and Build.xml templates (#706)
* Add Makefile hlc template

* Add hxcpp Build.xml hlc template

* Automate Makefile and Build.xml template builds

* Fix header file check when hlgen.makefilepath is set

* Avoid duplicate slashes in paths in hlc templates

* Avoid passing -std=c11 on MSVC in build.xml

* Add HASHLINK variable based paths in Build.xml

* Support debug builds with Build.xml template

* Improve makefile template

Use LDLIBS for passing libraries to the linker:
https://www.gnu.org/software/make/manual/html_node/Implicit-Variables.html#index-LDLIBS

Compile .c files into separate .o files to avoid recompiling the entire
project every time. Use .d files to keep track of .c files that
depend on .h files. For reference:
https://stackoverflow.com/a/52036564

* Add warning about auto generated makefile

* Add mingw support for makefile template

* Fix include path in generated makefile for mingw

For some reason, mingw doesn't like `-I./`, this causes it to be unable
to find the includes

* Keep makefile tmp files within hlgen.makefilepath

Previously it would always generate them next to the .c files,
regardless of whether hlgen.makefilepath had been set. Now it ensures
that it doesn't write outside of the Makefile directory.

This means that it first has to create the directories, otherwise the
compiler complains about directories not existing when it tries to
write the output files.

".SUFFIXES" also has to be set as an empty target, otherwise the built-
in rules mess up the rule dependencies.

* Allow HASHLINK for makefile template on linux/mac

* Simplify mingw flags for makefile template

* Let vs2019, vs2022 template build with makefile provided

* Makefile: generate separate linking flags for mac

* Add hlgen.makefile suggestion based on systemName

---------

Co-authored-by: Yuxiao Mao <yuxiao.mao@outlook.com>