[cmake] Improve support for android build (#798)
* [cmake] Add BUILD_SHARED_LIBS option

This is a standard cmake option for toggling between shared/static
library builds:

https://cmake.org/cmake/help/latest/variable/BUILD_SHARED_LIBS.html

* [cmake] Improve support for android build

CMakeLists.txt now fetches the required dependencies for android,
allowing for a more convenient build.

* [sdl] Fix compatibility with opengl es

* [sdl] Use gl es 3.2 header on android

This is required to support the following functions:
- glColorMaski (https://registry.khronos.org/OpenGL-Refpages/es3/html/glColorMask.xhtml)
- glGetProgramResourceIndex (https://registry.khronos.org/OpenGL-Refpages/es3/html/glGetProgramResourceIndex.xhtml)

* [hlc] Fix hlc_main.c for android sdks < 33

The functions in execinfo.h are only available with android sdk 33 and
above.

* [ci] Add android build

minSdkVersion/ANDROID_PLATFORM 24 is required by libuv:

https://github.com/libuv/libuv/blob/6cf854c11b1a6c5b629282ac334097ac8e6abb9e/SUPPORTED_PLATFORMS.md?plain=1#L13