[cmake] Link with shared openal again on android (#801)
* [cmake] Link with shared openal again on android

openal-soft has an LGPL license, which comes with some additional
restrictions which are easier to fulfil when linking dynamically.

https://www.gnu.org/licenses/gpl-faq.en.html#LGPLStaticVsDynamic

It is still useful to have the OPENAL_STATIC flag, in case someone wants
to link statically and is able to comply with the restrictions.

* [cmake] Support WITH_STATIC_OPENAL flag

This allows static linking to openal soft