Move windows.h from hl.h to hlsystem.h
`windows.h` is a big include that adds a lot of declarations and macros,
which may conflict with user code when `hl.h` is included in hlc code.

The last function that `hl.h` was using from this header was
`IsDebuggerPresent`, which is no longer called directly, so we can just
move use of this header out from `hl.h` and into a private `hlsystem.h`
header.