Remove windows.h include from hl.h (#804)
* Use hl_detect_debugger in hl.h for MSVC builds

This is already used on other platforms. It is also implemented on
windows using `IsDebuggerPresent` in src/std/error.c

* Move windows.h from hl.h to hlsystem.h

`windows.h` is a big include that adds a lot of declarations and macros,
which may conflict with user code when `hl.h` is included in hlc code.

The last function that `hl.h` was using from this header was
`IsDebuggerPresent`, which is no longer called directly, so we can just
move use of this header out from `hl.h` and into a private `hlsystem.h`
header.

* Use WIN32_LEAN_AND_MEAN for windows.h in hlc_main

This reduces the impact of the windows.h header on hlc code

* Remove windows _WINSOCKAPI_ hack

Now that hl.h no longer includes windows.h, this hack to prevent it
from including winsock.h is no longer necessary.

* Fix undefined strlen in src/std/date.c