Fix controller hotplug
SDL3 no longer uses controller index, but a joystick ID now. These IDs are not necessarily sequential, so controllers would often not work unless they were plugged in at launch, or may not work at all if the ids are higher than the count.

This requires a small heaps change as well.