Fix mingw build (#831)
* Fix mingw build errors

* Use SetThreadDescription for thread name on mingw

Mingw does not support __try and __except, so we cannot use the old
exception method for setting a thread name. We can however use
SetThreadDescription, which is the modern way of setting thread names.

* [make] Fix mingw build errors

* [cmake] Fix mingw build errors

* [ci] Add configuration for mingw cmake

* [ci] Use windows-2025 for mingw build

2022 has an old version of mingw which doesn't have SetThreadDescription

* [ui] Fix decrement on rvalue error with mingw

* [ci] Enable ui.hdll build with mingw

* [dx] Add missing cast in gamecontroller

Required by mingw compiler

* [sdl,dx] Fix include paths for mingw cross compile

* [sdl] Use system glext.h on mingw builds

glext.h is vendored for windows since that is not part of the windows
sdk. However, the vendored version is outdated and has issues on mingw,
and on mingw the header is provided by the system so that can be used
instead.

* [ci] Enable fmt, sdl, openal, directx mingw builds

* [video] Add missing cast for mingw

* [ci] Enable video.hdll build on mingw

* [ci] Use parallel build for cmake mingw

* [ci] Add mingw binaries to automated releases

---------

Co-authored-by: N3rdL0rd <168280212+N3rdL0rd@users.noreply.github.com>