Fix errors due to DOWNLOAD_EXTRACT_TIMESTAMP
`DOWNLOAD_EXTRACT_TIMESTAMP true` was added to suppress a warning on
cmake 3.24 and above. However, on older versions (like 3.22) this option
is not available, so it is parsed as part of BUILD_BYPRODUCTS or
URL_HASH, which results in errors.

Aside from this, setting the option to true is against the
recommendation of the docs, which recommend false for most cases
(despite the warning message):
https://cmake.org/cmake/help/latest/module/ExternalProject.html#url

Instead, we can enable the 3.24 policies globally when the version is
sufficient.

The warning is described on the page:
https://cmake.org/cmake/help/latest/policy/CMP0135.html